Изменил отображение счета и формат хранения счета

+++ database/migrations/2024_02_22_103243_create_wallets_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('type');
             $table->string('currency');
-            $table->float('balance');
+            $table->unsignedBigInteger('balance');
             $table->timestamps();
         });
     }
